backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
transformers>=4.41.0
torch>=2.2.0
accelerate>=0.30.0
python-dotenv>=1.0.1

# LLM providers (install what you use; safe to install all)
openai>=1.40.0
anthropic>=0.34.0
google-genai>=0.3.0